jooyparklove

+++ vesion3final_get/get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: jaeyojun <jaeyojun@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/31 12:35:53 by jaeyojun          #+#    #+#             */
-/*   Updated: 2023/04/05 23:50:47 by jaeyojun         ###   ########seoul.kr  */
+/*   Updated: 2023/04/06 16:35:11 by jaeyojun         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,6 @@
 
 # include <unistd.h>
 # include <stdlib.h>
-# include <string.h>
 
 # ifndef BUFFER_SIZE
 #  define BUFFER_SIZE 42
@@ -26,10 +25,9 @@ char		*ft_strdup(char *string);
 char		*ft_strjoin(char *s1, char *s2, int s2_len, int count);
 char		*ft_strcat(char *buff, char *prev);
 char		*ft_strchr(const char *s, int c);
-size_t		ft_strlen(const char *s);
-char		*ft_substr(char *s, unsigned int start, size_t len);
-char		*ft_substrab(char *s, unsigned int start, size_t len);
-
+int			ft_strlen(const char *s);
+char		*ft_substr(char *s, unsigned int start, unsigned int len);
+char		*ft_substrab(char *s, unsigned int start, int len);
 char		*ft_nonewline(char *line);
 char		*ft_readline(int fd, char *buff, char *backup);
 char		*get_next_line(int fd);
