+++ vesion3pr/get_next_line_bonus.h
@@ -6,7 +6,7 @@
 /*   By: jaeyojun <jaeyojun@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/31 12:35:53 by jaeyojun          #+#    #+#             */
-/*   Updated: 2023/04/06 16:34:37 by jaeyojun         ###   ########seoul.kr  */
+/*   Updated: 2023/04/06 15:52:15 by jaeyojun         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include <unistd.h>
 # include <stdlib.h>
+# include <string.h>
 
 # ifndef BUFFER_SIZE
 #  define BUFFER_SIZE 42
@@ -27,9 +28,9 @@ char		*ft_strdup(char *string);
 char		*ft_strjoin(char *s1, char *s2, int s2_len, int count);
 char		*ft_strcat(char *buff, char *prev);
 char		*ft_strchr(const char *s, int c);
-int			ft_strlen(const char *s);
-char		*ft_substr(char *s, unsigned int start, unsigned int len);
-char		*ft_substrab(char *s, unsigned int start, int len);
+size_t		ft_strlen(const char *s);
+char		*ft_substr(char *s, unsigned int start, size_t len);
+char		*ft_substrab(char *s, unsigned int start, size_t len);
 char		*ft_nonewline(char *line);
 char		*ft_readline(int fd, char *buff, char *backup);
 char		*get_next_line(int fd);
